+++ app/src/main/kotlin/ac/mdiq/podcini/net/sync/wifi/WifiSyncService.kt
@@ -261,7 +261,7 @@ import kotlin.math.min
                     .position(media.getPosition() / 1000)
                     .playedDuration(media.playedDuration / 1000)
                     .total(media.getDuration() / 1000)
-                    .isFavorite(item.isFavorite)
+                    .isFavorite(item.isSUPER)
                     .playState(item.playState)
                     .build()
                 queuedEpisodeActions.add(played)
@@ -334,7 +334,7 @@ import kotlin.math.min
                 it.media!!.setPosition(action.position * 1000)
                 it.media!!.playedDuration = action.playedDuration * 1000
                 it.media!!.setLastPlayedTime(action.timestamp!!.time)
-                it.rating = if (action.isFavorite) Rating.FAVORITE.code else Rating.UNRATED.code
+                it.rating = if (action.isFavorite) Rating.SUPER.code else Rating.UNRATED.code
                 it.playState = action.playState
                 if (hasAlmostEnded(it.media!!)) {
                     Logd(TAG, "Marking as played")