+++ app/src/main/kotlin/ac/mdiq/podcini/storage/model/FeedFilter.kt
@@ -51,9 +51,9 @@ class FeedFilter(vararg properties_: String) : Serializable {
         if (properties.contains(States.unrated.name)) ratingQuerys.add(" rating == ${Rating.UNRATED.code} ")
         if (properties.contains(States.trash.name)) ratingQuerys.add(" rating == ${Rating.TRASH.code} ")
         if (properties.contains(States.bad.name)) ratingQuerys.add(" rating == ${Rating.BAD.code} ")
-        if (properties.contains(States.neutral.name)) ratingQuerys.add(" rating == ${Rating.NEUTRAL.code} ")
+        if (properties.contains(States.OK.name)) ratingQuerys.add(" rating == ${Rating.OK.code} ")
         if (properties.contains(States.good.name)) ratingQuerys.add(" rating == ${Rating.GOOD.code} ")
-        if (properties.contains(States.favorite.name)) ratingQuerys.add(" rating == ${Rating.FAVORITE.code} ")
+        if (properties.contains(States.Super.name)) ratingQuerys.add(" rating == ${Rating.SUPER.code} ")
         if (ratingQuerys.isNotEmpty()) {
             val query = StringBuilder(" (" + ratingQuerys[0])
             if (ratingQuerys.size > 1) for (r in ratingQuerys.subList(1, ratingQuerys.size)) {
@@ -120,9 +120,9 @@ class FeedFilter(vararg properties_: String) : Serializable {
         unrated,
         trash,
         bad,
-        neutral,
+        OK,
         good,
-        favorite,
+        Super,
     }
 
     enum class FeedFilterGroup(val nameRes: Int, vararg values_: ItemProperties) {
@@ -131,9 +131,9 @@ class FeedFilter(vararg properties_: String) : Serializable {
         RATING(R.string.rating_label, ItemProperties(R.string.unrated, States.unrated.name),
             ItemProperties(R.string.trash, States.trash.name),
             ItemProperties(R.string.bad, States.bad.name),
-            ItemProperties(R.string.neutral, States.neutral.name),
+            ItemProperties(R.string.OK, States.OK.name),
             ItemProperties(R.string.good, States.good.name),
-            ItemProperties(R.string.favorite, States.favorite.name),
+            ItemProperties(R.string.Super, States.Super.name),
         ),
         HAS_VIDEO(R.string.has_video, ItemProperties(R.string.yes, States.has_video.name), ItemProperties(R.string.no, States.no_video.name)),
         PLAY_SPEED(R.string.play_speed, ItemProperties(R.string.global_speed, States.global_playSpeed.name), ItemProperties(R.string.custom_speed, States.custom_playSpeed.name)),